たぶん一旦全部できた

+++ service/user.go
@@ -41,7 +41,7 @@ func (s *UserService) ReadUser(ctx context.Context, username string, password st
 	return &res, nil
 }
 
-func (s *UserService) ReadUserWithId(userId int) (*model.ReadUserWithIdResponse, error){
+func (s *UserService) ReadUserWithId(userId int64) (*model.ReadUserWithIdResponse, error){
 	const query = `select username from users where id = ?`
 	var user model.ReadUserWithIdResponse
